cleaned the countries data
I hate CSV. I moved all the data around in r to get rid of columns that barely had any data, and I
wrote a python script that would take the csv and convert it to json because I didn't know how to do
it with r. We should be all set now with clean enough data!

I didn't want to include countries that had too many missing values, and I found that the sweetspot was around 8.