made big changes to the layout of the Game Input Component
I moved a lot of the code around and put them in different functions to make it as readable as possible.

I also completely ditched the csv and it now imports from a json file so it isn't asynchronous anymore, so that's a big plus.

I didn't change the game really at all, but the code is very different and totally refactored because of the csv -> json transition.